Deleted a bunch of ancillary classes. Added an output log class. The "ans" keyword is now working (to access the return value from the last computation)

+++ PhysicsEngine/Compiler/Token.cs
@@ -8,13 +8,13 @@ namespace PhysicsEngine {
 		private static readonly HashSet<string> functionLibrary = new HashSet<string>() { "Print", "Add", "Sum", "Derivative", "Integral" };
 		public TokenType TokenType;
 		public string TokenString;
-		public int TokenNumValue;
+		public double TokenNumValue;
 
 		public Token(string tokenString, TokenType tokenType) {
 			TokenString = tokenString;
 			TokenType = tokenType;
 			if (TokenType == TokenType.number) {
-				TokenNumValue = int.Parse(TokenString);
+				TokenNumValue = double.Parse(TokenString);
 			}
 			if (TokenType == TokenType.charString) {
 				if (functionLibrary.Contains(TokenString)) {
